feat(ArtyHub-Mislata/sprints#3): capa persistencia casi terminada

+++ src/main/java/es/artyhub/tienda_back/controller/webmodel/response/ArtworkDetailResponse.java
@@ -1,11 +1,14 @@
 package es.artyhub.tienda_back.controller.webmodel.response;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
+
 public record ArtworkDetailResponse(
     Long id,
     String name,
     String description,
     String imageUrl,
-    double price,
+    BigDecimal price,
     int stock,
     CategoryDetailResponse categoryDetailResponse
 ) {
