+++ src/main/java/es/artyhub/tienda_back/domain/dto/ArtworkDto.java
@@ -4,6 +4,9 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
+
 public record ArtworkDto(
 
         Long id,
@@ -19,7 +22,7 @@ public record ArtworkDto(
 
         @NotBlank(message = "El precio no puede ser vacío")
         @Positive(message = "El precio debe ser mayor a 0")
-        Double price,
+        BigDecimal price,
 
         @NotBlank(message = "El stock no puede ser vacío")
         @Positive(message = "El stock debe ser mayor a 0")
