+++ src/main/java/es/artyhub/tienda_back/domain/model/Artwork.java
@@ -2,20 +2,23 @@ package es.artyhub.tienda_back.domain.model;
 
 import es.artyhub.tienda_back.domain.dto.CategoryDto;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
+
 public class Artwork {
 
     private Long id;
     private String name;
     private String description;
     private String imageUrl;
-    private double price;
+    private BigDecimal price;
     private int stock;
     private CategoryDto categoryDto;
 
     public Artwork() {
     }
 
-    public Artwork(Long id, String name, String description, String imageUrl, double price, int stock,
+    public Artwork(Long id, String name, String description, String imageUrl, BigDecimal price, int stock,
             CategoryDto categoryDto) {
         this.id = id;
         this.name = name;
@@ -58,11 +61,11 @@ public class Artwork {
         this.imageUrl = imageUrl;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
